app/requirements.txt
django==4.1.7 
djangorestframework==3.14.0
pandas==2.0.3
psycopg2-binary==2.9.5
pytest-django==4.5.2
pytest==7.2.2
requests==2.31.0